+++ test/datalevin/search_test.clj
@@ -2,12 +2,10 @@
   (:require [datalevin.search :as sut]
             [datalevin.lmdb :as l]
             [datalevin.sparselist :as sl]
-            [datalevin.constants :as c]
             [datalevin.util :as u]
             [clojure.string :as s]
             [clojure.test :refer [is deftest testing]])
-  (:import [java.util UUID Map ArrayList]
-           [org.roaringbitmap RoaringBitmap]
+  (:import [java.util UUID ]
            [datalevin.sparselist SparseIntArrayList]
            [datalevin.search SearchEngine IndexWriter]))
 
@@ -143,6 +141,48 @@
     (is (empty? (sut/search engine "solar wind")))
     (is (= (sut/search engine "solar cap" {:display :offsets})
            [[:doc4 [["cap" [51]]]]]))
+    (l/close-kv lmdb)))
+
+(deftest search-143-test
+  (let [lmdb   (l/open-kv (u/tmp-dir (str "search-143-" (UUID/randomUUID))))
+        engine ^SearchEngine (sut/new-search-engine lmdb)]
+
+    (sut/add-doc engine 1 "a tent")
+    (sut/add-doc engine 2 "tent")
+
+    (is (= (sut/doc-count engine) 2))
+    (is (= (sut/doc-refs engine) [1 2]))
+
+    (let [[tid mw ^SparseIntArrayList sl]
+          (l/get-value lmdb (.-terms-dbi engine) "tent" :string :term-info true)]
+      (is (= (l/range-count lmdb (.-terms-dbi engine) [:all] :string) 1))
+      (is (= (l/get-value lmdb (.-terms-dbi engine) "tent" :string :int) tid))
+      (is (= mw 1.0))
+
+      (is (sl/contains-index? sl 1))
+      (is (= (sl/size sl) 2))
+      (is (= (seq (.-indices sl)) [1 2]))
+
+      (is (= (l/list-count lmdb (.-positions-dbi engine) [tid 1] :int-int)
+             (sl/get sl 1)
+             1))
+      (is (= (l/list-count lmdb (.-positions-dbi engine) [tid 2] :int-int)
+             (sl/get sl 2)
+             1))
+
+      (is (= (l/list-count lmdb (.-positions-dbi engine) [tid 3] :int-int)
+             0))
+      (is (nil? (sl/get sl 3)))
+
+      (is (= (l/get-list lmdb (.-positions-dbi engine) [tid 1] :int-int :int-int)
+             [[1 2]]))
+
+      (is (= (l/get-value lmdb (.-docs-dbi engine) 1 :int :doc-info true) [1 1]))
+      (is (= (l/get-value lmdb (.-docs-dbi engine) 2 :int :doc-info true) [1 2]))
+      (is (= (l/range-count lmdb (.-docs-dbi engine) [:all]) 2))
+      )
+
+    (is (= (sut/search engine "tent") [2 1]))
     (l/close-kv lmdb)))
 
 (deftest multi-domains-test